# Conflicts:
#	DuckyData1.0.0Alpha/Controllers/MessageController.cs
#	DuckyData1.0.0Alpha/Views/Account/Login.cshtml
#	DuckyData1.0.0Alpha/Views/Account/ResetPassword.cshtml
#	DuckyData1.0.0Alpha/Views/Home/Index.cshtml
#	DuckyData1.0.0Alpha/Views/Message/Create.cshtml
#	DuckyData1.0.0Alpha/Views/Shared/_Layout.cshtml
#	DuckyData1.0.0Alpha/Web.config

+++ DuckyData1.0.0Alpha/Controllers/MessageController.cs
@@ -1,64 +1,33 @@
 ﻿using AutoMapper;
+using DuckyData1._0._0Alpha;
+using DuckyData1._0._0Alpha.Models;
 using DuckyData1._0._0Alpha.ViewModels;
+using Microsoft.AspNet.Identity;
+using Microsoft.AspNet.Identity.EntityFramework;
 using System;
 using System.Collections.Generic;
-using System.Text.RegularExpressions;
-using System.Web.Mvc;
-using System.Web.Script.Serialization;
-using System.Web.Script.Services;
-using System.Web.Services;
 using System.Linq;
+using System.Web;
+using ACRCloud;
+using static System.Net.WebRequestMethods;
+using System.IO;
+using System.Web.Mvc;
+using DuckyData1._0._0Alpha.Factory.Account;
+using System.Data.Entity.Validation;
 
 namespace DuckyData1._0._0Alpha.Controllers
 {
     public class MessageController : Controller
     {
         private Manager m = new Manager();
+        private AccountFactory ac = new AccountFactory();
+        static private string uID = System.Web.HttpContext.Current.User.Identity.GetUserId();
 
-        
         // GET: Message
         [Authorize(Roles = "Admin")]
         public ActionResult Index()
         {
             return View(m.AllMsg());
-        }
-
-        [HttpPost]
-        private string SearchRecipients(string searchTerm, int sizelimit = 10)
-        {
-            Regex regexEmail = new Regex(@"^.*[\.].*[@]?.*$");
-            Match matchEmail = regexEmail.Match(searchTerm);
-
-            List<Models.Message> results = new List<Models.Message>();
-            var query = m.GetMessages();
-
-            if (matchEmail.Success == true)
-            {
-                query = query.Where(x => x.Recipient.ToUpper().Contains(searchTerm.ToUpper()))
-                            .GroupBy(x => x.Recipient)
-                            .Select(s => s.FirstOrDefault());
-            }
-
-            results = query.ToList();
-
-            List<Dictionary<string, string>> suggestions = new List<Dictionary<string, string>>(results.Count);
-            foreach (var result in results)
-            {
-                Dictionary<string, string> record = new Dictionary<string, string>();
-                string email = result.Recipient;
-                record.Add("email", email);
-
-                suggestions.Add(record);
-            }
-
-            return (new JavaScriptSerializer()).Serialize(suggestions);
-        }
-
-        [WebMethod]
-        [ScriptMethod(ResponseFormat = ResponseFormat.Json)]
-        public string GetRecipientSuggestions(string searchTerm)
-        {
-            return SearchRecipients(searchTerm);
         }
 
         // GET: Message
@@ -115,7 +84,6 @@ namespace DuckyData1._0._0Alpha.Controllers
         {
             if (ModelState.IsValid)
             {
-
                 var addedItem = m.SendMessage(newItem);
                 return RedirectToAction("Inbox", "Message");
             }
@@ -137,9 +105,14 @@ namespace DuckyData1._0._0Alpha.Controllers
         [HttpPost]
         public ActionResult Create(MessageAdd newItem)
         {
+            var user = ac.findUserById(uID);
+            if (user.gagged)
+            {
+                ModelState.AddModelError("SentDate", "Please select an option");
+                return View();
+            }
             if (ModelState.IsValid)
             {
-
                 var addedItem = m.SendMessage(newItem);
                 return RedirectToAction("Inbox", "Message");
             }
@@ -158,10 +131,8 @@ namespace DuckyData1._0._0Alpha.Controllers
             {
                 return HttpNotFound();
             }
-            if (DateTime.Compare(fetchedObject.SentDate, DateTime.Now) > 300)
-            {
-                return RedirectToAction("Inbox", "Message");
-            }
+            //if (DateTime.Compare(fetchedObject.SentDate, DateTime.Now) > 300)
+            
             return View(Mapper.Map<MessageEditForm>(fetchedObject));
             
         }
@@ -228,7 +199,10 @@ namespace DuckyData1._0._0Alpha.Controllers
         {
             if (!id.HasValue) { return HttpNotFound(); }
 
-
+            if( uID != m.GetMessageById(id.Value).UserId && !User.IsInRole("Admin"))
+            {
+                return RedirectToAction("Inbox", "Message");
+            }
             // Attempt to delete the item
             m.DeleteMessage(id.Value);
